refactor(core): extract createPromptModeHandlers
Pull the 5 prompt-mode UI handlers (submit/cancel/confirmDismiss/
cancelDismiss/toggleExpand) out of init() into a focused module.
Each handler is a small bookkeeping glue between the store actions,
the activation lifecycle, the copy orchestrator, and the prompt-mode
helpers; concentrating them here keeps init() free of the
'submit -> copy' / 'cancel -> dismiss-or-deactivate' branching.

index.tsx: 2217 -> 2171 (-46). 21 prompt-mode specs green.
